refactor(ors_control_app): slice FunctionTester into helpers + ResultMap
Phase 4b. Slice FunctionTester.tsx into the existing function-tester/
folder alongside samplePoints.ts:
- function-tester/helpers.tsx: RegionOption, GeoData, OptimizationStop,
  OptimizationVehicle, OptimizationParsed types; CARTO_LIGHT,
  OPTIMIZATION_PALETTE, PROFILE_LABELS, FUNCTIONS constants;
  cartoBasemap, bboxCenter, offsetPoint, isoRangeFor, isProvisionedRegion,
  generateSql, tryParseJson, decodePolyline, extractGeoData,
  parseMatrixResult, parseOptimizationResult, travelTimeColor,
  parseIsochroneOrigin pure helpers.
- function-tester/ResultMap.tsx: 320-line ResultMap visualisation
  component (deck.gl layers for directions / isochrones / matrix /
  optimization).

FunctionTester.tsx is now 436 lines (was 1,128) and only contains the
page-level state machine, form, and orchestration. Future commits can
split per-function panels (DirectionsPanel, IsochronesPanel, etc.) out
of the remaining body if desired.

Phase 4b of holistic-control-app-refactor-final.
No behaviour change. tsc --noEmit passes cleanly.
